Fix imitation learning algorithms and align environments with Gymnasium v5
  Algorithms:
  - Use the updated discriminator state for rewards in GAIL, AMP, AIRL
  - Fix TRIRL-DTRL imports to flax_full_jit
  - Add nr_epochs_disc and drop the retraining read in AIRL
  - Add handle_absorbing_states and H_terminal in NEAR
  - Shape the AIRL discriminator init log probability as one per sample
  - Anneal the discriminator from learning_rate_disc
  - Pass maximum_eta unnegated in TRIRL-DTRL
  - Drop the minibatch remainder before reshaping; reject a minibatch
    larger than the batch
  - Sample expert batches with replacement
  - Cast expert data to float32 and derive num_data_samples from it
  - Store the discriminator in GAIL, AMP and TRIRL checkpoints, and the
    reward network or the buffers in TRIRL
  - Rename the model archive in AIRL and NEAR
  - Remove the unused dim attribute

  Environments:
  - Compute the forward reward from displacement over the frame-skipped
    step, using torso xpos for Ant, mass center for Humanoid, qpos[0] for
    the rest
  - Scale the healthy reward by is_healthy
  - Add the state finiteness check and inclusive height bounds to Ant, and
    the healthy state range to Hopper
  - Compute the Ant contact cost over the full contact force array
  - Add a dt attribute outside the render branch
  - Add center and scale to BoxSpace
  - Remove an unused import

  Documentation:
  - Correct the NEAR name in the documentation index
  - Describe Point Maze as a custom task

+++ rl_x/algorithms/gail_ppo/flax_full_jit/data_utils.py
@@ -8,13 +8,13 @@ def prepare_expert_data(data_path, cutoff=1):
     expert_files = np.load(data_path)
 
     def _flatten_feature_array(x):
-        x = np.asarray(x)
+        x = np.asarray(x, dtype=np.float32)
         if x.ndim <= 2:
             return x
         return x.reshape(-1, x.shape[-1])
 
     def _flatten_scalar_array(x):
-        return np.asarray(x).reshape(-1)
+        return np.asarray(x, dtype=np.float32).reshape(-1)
 
     states = _flatten_feature_array(expert_files["states"])
     actions = _flatten_feature_array(expert_files["actions"])
